APEXMALHAR-2056 #resolve #comment moved Serde interface.
