+++ library/src/main/java/org/apache/apex/malhar/lib/spillable/SpillableComplexComponent.java
@@ -19,9 +19,11 @@
 package org.apache.apex.malhar.lib.spillable;
 
 import org.apache.apex.malhar.lib.spillable.Spillable.SpillableComponent;
+import org.apache.apex.malhar.lib.utils.Serde;
 
 import com.datatorrent.api.Component;
 import com.datatorrent.api.Context.OperatorContext;
+
 
 /**
  * This is a composite component containing spillable data structures. This should be used as
